Move remaining util functions to another module
